make json serialization test a global helper function
